Extract resource and medium loading
