+++ app/models/csv_dumps/finds_dump_scope.rb
@@ -1,0 +1,7 @@
+module CsvDumps
+  class FindsDumpScope
+    def self.find(resource_type, resource_id)
+      resource_type.camelize.constantize.find(resource_id)
+    end
+  end
+end
